cancel timeout watches to conserve resources
When an action that is guarded by a timeout successfully
completes, the timout watch used to keep running. That
is just a waste of time and resources.

This commit makes sure that when a timeout action finishes,
the timeout watch is cancelled. It also significantly improves
test coverage for timeouts.